requirements.txt
numpy>=1.12
pandas
scipy
spacy
image-search
opencv-python
grequests
smallutils
gevent
reportlab
Pillow